requirements.txt
langchain==1.2.15
langchain-core==1.2.28
langchain-openai==1.1.12
langchain-community==0.4.1
langgraph==1.1.6
langgraph-prebuilt==1.0.9
pydantic==2.13.0
python-dotenv==1.2.2
yfinance==1.3.0
tavily-python==0.7.23
requests==2.33.1
streamlit==1.56.0
